+++ staek-jdbc/src/main/java/com/staekframework/test/User/User.java
@@ -1,4 +1,4 @@
-package com.staekframework.test;
+package com.staekframework.test.User;
 
 public class User {
